requirements.txt
torch
torchvision
timm
opencv-python
numpy
matplotlib
streamlit
scikit-learn
grad-cam
Pillow